+++ src/GAT2VEC/paths.py
